% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as_data_frame}
\alias{as_data_frame}
\alias{as.data.frame.ouchtree}
\alias{as.data.frame.brown}
\alias{as.data.frame.hansentree}
\title{Coerce an \pkg{ouch} object to a data frame}
\usage{
\method{as.data.frame}{ouchtree}(x, ...)

\method{as.data.frame}{brown}(x, ...)

\method{as.data.frame}{hansentree}(x, ...)
}
\arguments{
\item{x}{any \R object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Coerce an \pkg{ouch} object to a data frame
}
\seealso{
Other methods for ouch trees: 
\code{\link{bootstrap}()},
\code{\link{coef}()},
\code{\link{logLik}},
\code{\link{ouch-package}},
\code{\link{paint}()},
\code{\link{plot}()},
\code{\link{print}()},
\code{\link{simulate}()},
\code{\link{summary}()},
\code{\link{update}()}
}
\concept{methods for ouch trees}
