% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap,browntree-method}
\alias{bootstrap,hansentree-method}
\alias{bootstrap,missing-method}
\alias{bootstrap,ANY-method}
\title{Bootstrapping for uncertainty quantification}
\usage{
\S4method{bootstrap}{browntree}(object, nboot = 200, seed = NULL, ...)

\S4method{bootstrap}{hansentree}(object, nboot = 200, seed = NULL, ...)

\S4method{bootstrap}{missing}(object, ...)

\S4method{bootstrap}{ANY}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{nboot}{integer; number of bootstrap replicates.}

\item{seed}{integer; setting \code{seed} to a non-\code{NULL} value allows one to fix the random seed (see \code{\link[ouch:simulate]{simulate}}).}

\item{...}{Additional arguments are passed to \code{\link[ouch:update]{update}}.}
}
\description{
Generic bootstrapping for \pkg{ouch} models.
}
\details{
\code{bootstrap} performs a parametric bootstrap for estimation of confidence intervals.
}
