\name{hansentree}
\docType{class}
\title{hansen tree object}
\alias{hansentree}
\alias{hansentree-class}
\description{
  A fitted phylogenetic Hansen-model object.
}
\details{
  The function \code{hansen} creates a \code{hansentree} object by fitting a Hansen model to data.
}
\section{Methods}{
  See \link{hansentree-methods}.
}
\author{Aaron A. King kingaa at umich dot edu}
\seealso{\code{\link{ouchtree}}, \code{\link{hansen}}, \link{hansentree-methods}}
\keyword{models}
