% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brown.R, R/hansen.R, R/loglik.R
\name{logLik,browntree-method}
\alias{logLik,browntree-method}
\alias{logLik,hansentree-method}
\alias{loglik}
\title{Log likelihood of a fitted model}
\usage{
\S4method{logLik}{browntree}(object)

\S4method{logLik}{hansentree}(object)
}
\arguments{
\item{object}{fitted model object}
}
\value{
\code{logLik} returns a numeric value.
}
\description{
\code{logLik} extracts the log likelihood from a fitted model object.
}
