% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R, R/brown.R, R/hansen.R
\docType{methods}
\name{simulate}
\alias{simulate}
\alias{simulate,browntree-method}
\alias{simulate,hansentree-method}
\title{Simulations of a phylogenetic trait model}
\usage{
\S4method{simulate}{browntree}(object, nsim = 1, seed = NULL, ...)

\S4method{simulate}{hansentree}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{fitted model object}

\item{nsim}{integer; number of independent simulations.}

\item{seed}{integer; if non-\code{NULL}, the RNG will be initialized with this seed for the simulations.
The RNG will be reset to its pre-existing state when \code{simulate} returns.}

\item{...}{additional arguments, ignored.}
}
\value{
\code{simulate} returns a list of data-frames, each comparable to the original data.
}
\description{
\code{simulate} generates random deviates from a fitted model.
}
\seealso{
Other methods: 
\code{\link{bootstrap}()},
\code{\link{coef}()},
\code{\link{logLik}},
\code{\link{plot}()},
\code{\link{print}()},
\code{\link{summary}()},
\code{\link{update}()}
}
\concept{methods}
