% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R, R/brown.R, R/hansen.R
\name{summary}
\alias{summary}
\alias{summary,browntree-method}
\alias{summary,hansentree-method}
\alias{summary-hansentree,}
\title{Summary methods}
\usage{
\S4method{summary}{browntree}(object, ...)

\S4method{summary}{hansentree}(object, ...)
}
\arguments{
\item{object}{fitted model object.}

\item{...}{additional arguments, ignored.}
}
\value{
\code{summary} applied to a \code{browntree} object returns information about the fitted model, including parameter estimates and quantities describing the goodness of fit.

\code{summary} applied to a \code{hansentree} method displays the estimated \eqn{\alpha}{alpha} and \eqn{\sigma^2}{sigma^2} matrices as well as various quantities describing the goodness of model fit.
}
\description{
Summary methods
}
\seealso{
Other methods: 
\code{\link{bootstrap}()},
\code{\link{coef}()},
\code{\link{logLik}},
\code{\link{plot}()},
\code{\link{print}()},
\code{\link{simulate}()},
\code{\link{update}()}
}
\concept{methods}
