% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen.R
\name{hansen}
\alias{hansen}
\alias{hansentree-class}
\title{Ornstein-Uhlenbeck models of trait evolution}
\usage{
hansen(
  data,
  tree,
  regimes,
  sqrt.alpha,
  sigma,
  fit = TRUE,
  method = c("Nelder-Mead", "subplex", "BFGS", "L-BFGS-B"),
  hessian = FALSE,
  ...
)
}
\arguments{
\item{data}{Phenotypic data for extant species, i.e., species at the terminal twigs of the phylogenetic tree.
This can either be a single named numeric vector, a list of \code{nchar} named vectors, or a data frame containing \code{nchar} data variables.
There must be an entry per variable for every node in the tree; use \code{NA} to represent missing data.
If the
data are supplied as one or more named vectors, the names attributes are taken to correspond to the node names specified when the \code{ouchtree} was constructed (see \code{\link{ouchtree}}).
If the data are supplied as a
data-frame, the rownames serve that purpose.}

\item{tree}{A phylogenetic tree, specified as an \code{ouchtree} object.}

\item{regimes}{A vector of codes, one for each node in the tree, specifying the selective regimes hypothesized to have been operative.
Corresponding to each node, enter the code of the regime hypothesized for the branch segment terminating in that node.
For the root node, because it has no branch segment terminating on it, the regime specification is irrelevant.
If there are \code{nchar} quantitative characters, then one can specify a single set of \code{regimes} for all characters or a list of \code{nchar} regime specifications, one for each character.}

\item{sqrt.alpha, sigma}{These are used to initialize the optimization algorithm.
The selection strength matrix \eqn{\alpha}{alpha} and the random drift variance-covariance matrix \eqn{\sigma^2}{sigma^2} are parameterized by their matrix square roots.
Specifically, these initial guesses are each packed into lower-triangular matrices (column by column).
The product of this matrix with its transpose is the \eqn{\alpha}{alpha} or \eqn{\sigma^2}{sigma^2} matrix.
See Details for more information.}

\item{fit}{If \code{fit=TRUE}, then the likelihood will be maximized.
If \code{fit=FALSE}, the likelihood will be evaluated at the specified values of \code{sqrt.alpha} and \code{sigma};
the optima \code{theta} will be returned as well.}

\item{method}{The method to be used by the optimization algorithm.
See \code{\link[subplex:subplex]{subplex::subplex}} and \code{\link[stats:optim]{stats::optim}} for information on the available options.}

\item{hessian}{If \code{hessian=TRUE}, then the Hessian matrix will be computed by \code{optim}.}

\item{\dots}{Additional arguments will be passed as \code{control} options to \code{optim} or \code{subplex}.
See \code{\link[stats:optim]{stats::optim()}} and \code{\link[subplex:subplex]{subplex::subplex()}} for information on the available options.}
}
\value{
\code{hansen} returns an object of class \code{hansentree}.
}
\description{
The function \code{hansen} fits an Ornstein-Uhlenbeck model to data.
The fitting is done using \code{optim} or \code{subplex}.
}
\details{
The Hansen model for the evolution of a multivariate trait \eqn{X} along a lineage can be written as a stochastic differential equation (Ito diffusion)
\deqn{dX=\alpha(\theta(t)-X(t))dt+\sigma dB(t),}{dX = alpha (theta(t)-X(t)) dt + sigma dB(t),}
where \eqn{t} is time along the lineage,
\eqn{\theta(t)}{theta(t)} is the optimum trait value, \eqn{B(t)} is a standard Wiener process (Brownian motion),
and \eqn{\alpha}{alpha} and \eqn{\sigma}{sigma} are matrices
quantifying, respectively, the strength of selection and random drift.
Without loss of generality, one can assume \eqn{\sigma}{sigma} is lower-triangular.
This is because only the infinitesimal variance-covariance matrix
\eqn{\sigma^2=\sigma\sigma^T}{sigma^2 = sigma\%*\%transpose(sigma)}
is identifiable, and for any admissible variance-covariance matrix, we can choose \eqn{\sigma}{sigma} to be lower-triangular.
Moreover, if we view the basic model as describing evolution on a fitness landscape, then \eqn{\alpha}{alpha} will be symmetric.
If we further restrict ourselves to the case of stabilizing selection, \eqn{\alpha}{alpha} will be positive definite as well.
We make these assumptions and therefore can assume that the matrix \eqn{\alpha}{alpha} has a lower-triangular square root.

The \code{hansen} code uses unconstrained numerical optimization to maximize the likelihood.
To do this, it parameterizes the \eqn{\alpha}{alpha} and \eqn{\sigma^2}{sigma^2} matrices in a special way:
each matrix is parameterized by \code{nchar*(nchar+1)/2} parameters, where \code{nchar} is the number of quantitative characters.
Specifically, the parameters initialized by the \code{sqrt.alpha} argument of \code{hansen} are used
to fill the nonzero entries of a lower-triangular matrix (in column-major order),
which is then multiplied by its transpose to give the selection-strength matrix.
The parameters specified in \code{sigma} fill the nonzero entries in the lower triangular \eqn{\sigma}{sigma} matrix.
When \code{hansen} is executed, the numerical optimizer maximizes the likelihood over these parameters.
}
\examples{
## Analysis of sexual size dimorphism data
tree <- with(anolis.ssd,ouchtree(node,ancestor,time/max(time),species))
plot(tree,node.names=TRUE)

h1 <- brown(anolis.ssd['log.SSD'],tree)
h1
plot(h1)

h2 <- hansen(anolis.ssd['log.SSD'],tree,anolis.ssd['OU.1'],sqrt.alpha=1,sigma=1)
h2
plot(h2)

h3 <- hansen(anolis.ssd['log.SSD'],tree,anolis.ssd['OU.7'],sqrt.alpha=1,sigma=1)
h3
plot(h3)

### Darwin's finches.
### The data were taken from package 'geiger' due to the latter being orphaned.
if (requireNamespace("ape")) {

  data(geospiza)
  plot(geospiza$phy)
  print(geospiza$dat)
  
### make an ouchtree out of the phy-format tree
  ot <- ape2ouch(geospiza$phy)

### merge data with tree info
  otd <- as(ot,"data.frame")
  otd <- merge(otd,geospiza$dat,by.x="labels",by.y="row.names",all=TRUE)
### row-names are used by 'hansen'
  rownames(otd) <- otd$nodes
  print(otd)
### this data-frame now contains the data as well as the tree geometry

### now remake the ouch tree
  ot <- with(otd,ouchtree(nodes=nodes,ancestors=ancestors,times=times,labels=labels))
  plot(ot)

  b1 <- brown(tree=ot,data=otd[c("tarsusL","beakD")])
  summary(b1)

### evaluate an OU model with a single, global selective regime
  otd$regimes <- as.factor("global")
  h1 <- hansen(
    tree=ot,
    data=otd[c("tarsusL","beakD")],
    regimes=otd["regimes"],
    sqrt.alpha=c(1,0,1),
    sigma=c(1,0,1),
    maxit=10000
  )
  summary(h1)
  plot(h1)

}
}
\references{
\Hansen1997

\Butler2004

\Cressler2015
}
\seealso{
\code{\link[stats:optim]{stats::optim}}, \code{\link[subplex:subplex]{subplex::subplex}}, \code{\link{bimac}}, \code{\link{anolis.ssd}}

Other phylogenetic comparative models: 
\code{\link{brown}()},
\code{\link{ouch-package}},
\code{\link{ouchtree}},
\code{\link{paint}()}
}
\author{
Aaron A. King
}
\concept{phylogenetic comparative models}
\keyword{models}
