% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R, R/brown.R, R/hansen.R
\name{logLik}
\alias{logLik}
\alias{logLik,browntree-method}
\alias{logLik,hansentree-method}
\alias{logLik-hansentree,}
\title{Log likelihood of a fitted model}
\usage{
\S4method{logLik}{browntree}(object)

\S4method{logLik}{hansentree}(object)
}
\arguments{
\item{object}{fitted model object}
}
\value{
\code{logLik} returns a numeric value.
}
\description{
\code{logLik} extracts the log likelihood from a fitted model object.
}
\seealso{
Other methods: 
\code{\link{bootstrap}()},
\code{\link{coef}()},
\code{\link{plot}()},
\code{\link{print}()},
\code{\link{simulate}()},
\code{\link{summary}()},
\code{\link{update}()}
}
\concept{methods}
