% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/ouchtree.R, R/hansen.R
\name{plot}
\alias{plot}
\alias{plot,ouchtree-method}
\alias{plot,hansentree-method}
\title{Plotting functions.}
\usage{
\S4method{plot}{ouchtree}(x, y, regimes = NULL, node.names = FALSE, legend = TRUE, ..., labels)

\S4method{plot}{hansentree}(x, y, regimes, ...)
}
\arguments{
\item{x}{object to plot.}

\item{y}{ignored.}

\item{regimes}{factor or character; a vector of regime paintings.}

\item{node.names}{logical; should node names be displayed?}

\item{legend}{logical; display a legend?}

\item{...}{additional arguments, passed to \code{text}.}

\item{labels}{character; taxon labels.}
}
\description{
Plot phylogenies.
}
\seealso{
Other methods: 
\code{\link{bootstrap}()},
\code{\link{coef}()},
\code{\link{logLik}},
\code{\link{print}()},
\code{\link{simulate}()},
\code{\link{summary}()},
\code{\link{update}()}
}
\concept{methods}
