% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouchtree.R, R/brown.R, R/hansen.R, R/print.R
\name{print,ouchtree-method}
\alias{print,ouchtree-method}
\alias{show,ouchtree-method}
\alias{show,browntree-method}
\alias{print,browntree-method}
\alias{print,hansentree-method}
\alias{show,hansentree-method}
\alias{print}
\title{Print methods}
\usage{
\S4method{print}{ouchtree}(x, ...)

\S4method{show}{ouchtree}(object)

\S4method{show}{browntree}(object)

\S4method{print}{browntree}(x, ...)

\S4method{print}{hansentree}(x, ...)

\S4method{show}{hansentree}(object)
}
\arguments{
\item{...}{additional arguments, ignored.}

\item{object, x}{object to display.}
}
\value{
\code{print} displays the tree as a table, along with the coefficients of the fitted model and diagnostic information.

\code{print} displays the tree as a table, along with the coefficients of the fitted model and diagnostic information.

\code{print} displays the tree as a table, with (possibly) other information.
}
\description{
Print methods
}
